refactore: watchlist/screener widgets
